Use `_create_table` instead
